LayoutTests/platform/ios/fast/dom/HTMLImageElement/sizes/image-sizes-js-change-expected.txt

PASS currentSrcFileName("foo") is "image-set-4x.png"
PASS successfullyParsed is true

TEST COMPLETE

image-sizes-js-change.html has MIME type text/html
srcset-helper.js has MIME type text/javascript
js-test.js has MIME type text/javascript
image-set-4x.png has MIME type image/png
image-sizes-js-change.html has MIME type text/html
srcset-helper.js has MIME type text/javascript
js-test.js has MIME type text/javascript
image-set-4x.png has MIME type image/png
